requirements.txt
homeassistant==2025.1.4
pip>=21.3.1
ruff==0.9.6
pycryptodomex
pytest==8.3.4
